Averaging and smoothing

SPEDAS.smooth — Function

smooth(da::AbstractDimArray, window; dim=Ti, suffix="_smoothed", kwargs...)

Smooths a time series by computing a moving average over a sliding window.

The size of the sliding window can be either:

Quantity: A time duration that will be converted to number of samples based on data resolution

Integer: Number of samples directly

Arguments

dims=Ti: Dimension along which to perform smoothing (default: time dimension)

suffix="_smoothed": Suffix to append to the variable name in output

kwargs...: Additional arguments passed to RollingWindowArrays.rolling

Exponential smoothing

Wikipedia