Time Series Utilities

The following utilities are designed to simplify common data analysis tasks for time series data. While most are thin wrappers around existing Julia functions, they ensure proper handling of metadata and dimensions.

Most of the utilities operate on the time dimension by default, but you can specify other dimensions using the dim or query parameter.

Statistics

Missing docstring.

Missing docstring for tstat. Check Documenter's build log for details.

SPEDAS.tmean — Function

tmean(x; dim=nothing, query=nothing)

Calculate the arithmetic mean of x along the dim dimension.

It returns a value if x is a vector along the dim dimension, otherwise returns a DimArray with the specified dimension dropped.

If dim is not specified, it defaults to the query dimension (dimension of type TimeDim by default).

https://github.com/JuliaLang/julia/issues/54542

SPEDAS.tmedian — Function

tmedian(x; dim=nothing, query=nothing)

Calculate the median of x along the dim dimension.

It returns a value if x is a vector along the dim dimension, otherwise returns a DimArray with the specified dimension dropped.

If dim is not specified, it defaults to the query dimension (dimension of type TimeDim by default).

SPEDAS.tsum — Function

tsum(x; dim=nothing, query=nothing)

Calculate the sum of x along the dim dimension.

It returns a value if x is a vector along the dim dimension, otherwise returns a DimArray with the specified dimension dropped.

If dim is not specified, it defaults to the query dimension (dimension of type TimeDim by default).

SPEDAS.tvar — Function

tvar(x; dim=nothing, query=nothing)

Calculate the variance of x along the dim dimension.

It returns a value if x is a vector along the dim dimension, otherwise returns a DimArray with the specified dimension dropped.

If dim is not specified, it defaults to the query dimension (dimension of type TimeDim by default).

SPEDAS.tstd — Function

tstd(x; dim=nothing, query=nothing)

Calculate the standard deviation of x along the dim dimension.

It returns a value if x is a vector along the dim dimension, otherwise returns a DimArray with the specified dimension dropped.

If dim is not specified, it defaults to the query dimension (dimension of type TimeDim by default).

SPEDAS.tsem — Function

tsem(x; dim=nothing, query=nothing)

Calculate the standard error of the mean of x along the dim dimension.

It returns a value if x is a vector along the dim dimension, otherwise returns a DimArray with the specified dimension dropped.

If dim is not specified, it defaults to the query dimension (dimension of type TimeDim by default).

Arithmetic

SPEDAS.tcross — Function

tcross(x, y; dims=TimeDim, stack=nothing)

Compute the cross product of two (arrays of) vectors along the dims dimension.

References:

https://docs.xarray.dev/en/stable/generated/xarray.cross.html

SPEDAS.tderiv — Function

tderiv(data, times; dims = 1)

Compute the time derivative of data with respect to times.

tderiv(data; dims = Ti)

Compute the time derivative of data.

See also: deriv_data - PySPEDAS

SPEDAS.tdot — Function

tdot(x, y; dims=TimeDim)

Dot product of two arrays x and y along the dims dimension.

SPEDAS.tsubtract — Function

tsubtract(x, f=nanmedian; dims=timedim(x))

Subtract a statistic (default function f: nanmedian) along dimensions (default: time dimension) from x.

Time-Domain Operations

SPEDAS.tclip — Function

tclip(d, t0, t1; query=nothing, sort=false)

Clip a dimension or DimArray to a time range [t0, t1].

For unordered dimensions, the dimension should be sorted before clipping (see tsort).

SPEDAS.tview — Function

tview(d, t0, t1)

View a dimension or DimArray in time range [t0, t1].

SPEDAS.tmask — Function

tmask(da, args...; kwargs...)

Non-mutable version of tmask!. See also tmask!.

SPEDAS.tshift — Function

tshift(x; dim=TimeDim, t0=nothing, new_dim=nothing)

Shift the dim of x by t0.