Time History of Events and Macroscale Interactions during Substorms (THEMIS)

SPEDAS.THEMIS — Module

Sub-module for "Time History of Events and Macroscale Interactions during Substorms (THEMIS)"

To load project, project-specific instrument and dataset variables into scope:

using SPEDAS.THEMIS

Configuration File: THEMIS.toml

SPEDAS.THEMIS.themis — Constant

Project: Time History of Events and Macroscale Interactions during Substorms
  Metadata (Dict): 
    abbreviation: THEMIS
    links: ["https://en.wikipedia.org/wiki/THEMIS"]
    spase_id: spase://SMWG/Observatory/THEMIS

Instruments

Datasets